refactor: 디자인 패턴 변경 (#7)
atomic -> compound
