+++ src/components/common/Text.tsx
